Refactor 2 (#32)
* Split in ui/ux

* Refactor provider rpc and chain id
