+++ package.json
@@ -13,9 +13,9 @@
     "lint:fix": "npx eslint 'src/**/*.{js,ts,tsx}' 'test/**/*.{js,ts,tsx}' --ext .json,.js,.ts --fix",
     "build": "webpack --env=production",
     "build:dev": "webpack --env=development",
-    "sample:dapp": "concurrently \"node sample/launcher.js\" \"serve dist -l 3005\" \"serve sample/front -l 3006 \"",
-    "sample:open": "concurrently \"node sample/launcher.js backend\" \"serve dist -l 3005\" \"serve sample/front -l 3006 \" \"nodemon sample/back/index.js \"",
-    "sample:permissioned": "concurrently \"node sample/launcher.js backend\" \"serve dist -l 3005\" \"serve sample/front -l 3006 \" \"nodemon sample/back/index.js --permissioned \"",
+    "sample:dapp": "concurrently \"npm run build:dev\" \"node sample/launcher.js\" \"serve dist -l 3005\" \"serve sample/front -l 3006\"",
+    "sample:open": "concurrently \"npm run build:dev\" \"node sample/launcher.js backend\" \"serve dist -l 3005\" \"serve sample/front -l 3006\" \"nodemon sample/back/index.js\"",
+    "sample:permissioned": "concurrently \"node sample/launcher.js backend\" \"serve dist -l 3005\" \"serve sample/front -l 3006\" \"nodemon sample/back/index.js --permissioned\"",
     "serve": "serve dist -l 3005",
     "prepublish": "npm run build"
   },
