+++ sample/front/index.html
@@ -42,6 +42,8 @@
     <script src="https://unpkg.com/@walletconnect/web3-provider@1.2.1/dist/umd/index.min.js"></script>
 
     <script type="text/javascript">
+      window.ethereum.autoRefreshOnNetworkChange = false
+
       // detect flavor
       const withBackend = new URLSearchParams(window.location.search).get('backend') === 'yes'
       const backendUrl = withBackend && 'http://localhost:3007'
