--- src/components/index.tsx
@@ -1,1 +1,0 @@
-export { Core } from './Core'
